+++ volmdlr/edges.py
@@ -21,8 +21,9 @@ from matplotlib import __version__ as _mpl_version
 from mpl_toolkits.mplot3d import Axes3D
 from packaging import version
 
-import dessia_common as dc
 import plot_data.core as plot_data
+import dessia_common.core as dc
+import volmdlr.utils.intersections as vm_utils_intersections
 import volmdlr.core_compiled
 import volmdlr.core
 import volmdlr.geometry
@@ -58,6 +59,11 @@ def insert_knots_and_mutiplicity(knots, knot_mutiplicities, knot_to_add, num):
 
 
 class Edge(dc.DessiaObject):
+    """
+    Defines a simple edge Object.
+
+    """
+
     def __init__(self, start, end, name=''):
         self.start = start
         self.end = end
@@ -178,7 +184,8 @@ class Edge(dc.DessiaObject):
 
 class Line(dc.DessiaObject):
     """
-    Abstract class
+    Abstract class.
+
     """
 
     def __init__(self, point1, point2, name=''):
@@ -190,7 +197,7 @@ class Line(dc.DessiaObject):
     def __getitem__(self, key):
         if key == 0:
             return self.point1
-        elif key == 1:
+        if key == 1:
             return self.point2
         raise IndexError
 
@@ -229,8 +236,8 @@ class Line(dc.DessiaObject):
         """
         Sort point along a line.
 
-        :param points: list of points to be sorted
-        :return: sorted points
+        :param points: list of points to be sorted.
+        :return: sorted points.
         """
         return sorted(points, key=self.abscissa)
 
@@ -243,11 +250,11 @@ class Line(dc.DessiaObject):
         """
         Verifies if a line is between two points.
 
-        :param point1: first point
-        :type point1: volmdlr.Point2D
-        :param point2: second point
-        :type point2: volmdlr.Point2D
-        returns True is line is between the two given points or False if not
+        :param point1: first point.
+        :type point1: volmdlr.Point2D.
+        :param point2: second point.
+        :type point2: volmdlr.Point2D.
+        returns True is line is between the two given points or False if not.
         """
 
         if point1 == point2:
@@ -261,7 +268,8 @@ class Line(dc.DessiaObject):
 
 class LineSegment(Edge):
     """
-    Abstract class
+    Abstract class.
+
     """
 
     def length(self):
@@ -287,8 +295,8 @@ class LineSegment(Edge):
         Computes a unit direction vector for the line segment.
 
         :param abscissa: defines where in the line segement the unit
-            direction vector is to be calculated
-        :return: The unit direction vector of the LineSegement
+            direction vector is to be calculated.
+        :return: The unit direction vector of the LineSegement.
         """
         direction_vector = self.direction_vector()
         direction_vector.normalize()
@@ -297,8 +305,8 @@ class LineSegment(Edge):
     def direction_vector(self, abscissa=0.):
         """
         :param abscissa: defines where in the line segement
-            direction vector is to be calculated
-        :return: The direction vector of the LineSegement
+            direction vector is to be calculated.
+        :return: The direction vector of the LineSegement.
         """
         if not self._direction_vector:
             self._direction_vector = self.end - self.start
@@ -307,16 +315,16 @@ class LineSegment(Edge):
     def normal_vector(self, abscissa=0.):
         """
         :param abscissa: defines where in the line_segement
-        normal vector is to be calculated
-        :return: The normal vector of the LineSegement
+        normal vector is to be calculated.
+        :return: The normal vector of the LineSegement.
         """
         return self.direction_vector(abscissa).normal_vector()
 
     def unit_normal_vector(self, abscissa=0.):
         """
-        :param abscissa: defines where in the line_segement
-        unit normal vector is to be calculated
-        :return: The unit normal vector of the LineSegement
+
+        :param abscissa: defines where in the line_segement unit normal vector is to be calculated.
+        :return: The unit normal vector of the LineSegement.
         """
         return self.unit_direction_vector(abscissa).normal_vector()
 
@@ -345,12 +353,11 @@ class LineSegment(Edge):
 
 
 class BSplineCurve(Edge):
-    _non_serializable_attributes = ['curve']
     """
     An abstract class for B-spline curves. The following rule must be
-    respected : `number of knots = number of control points + degree + 1`
+    respected : `number of knots = number of control points + degree + 1`.
 
-    :param degree: The degree of the B-spline curve
+    :param degree: The degree of the B-spline curve.
     :type degree: int
     :param control_points: A list of 2 or 3 dimensional points
     :type control_points: Union[List[:class:`volmdlr.Point2D`],
@@ -368,6 +375,7 @@ class BSplineCurve(Edge):
     :param name: The name of the B-spline curve. Default value is ''
     :type name: str, optional
     """
+    _non_serializable_attributes = ['curve']
 
     def __init__(self,
                  degree: int,
@@ -419,8 +427,8 @@ class BSplineCurve(Edge):
         """
         Reverses the B-spline's direction by reversing its control points.
 
-        :return: A reversed B-spline curve
-        :rtype: :class:`volmdlr.edges.BSplineCurve`
+        :return: A reversed B-spline curve.
+        :rtype: :class:`volmdlr.edges.BSplineCurve`.
         """
         return self.__class__(
             degree=self.degree,
@@ -453,9 +461,9 @@ class BSplineCurve(Edge):
 
     def length(self):
         """
-        Returns the length of the B-spline curve
+        Returns the length of the B-spline curve.
 
-        :return: The length of the B-spline curve
+        :return: The length of the B-spline curve.
         :rtype: float
         """
         if not self._length:
@@ -492,11 +500,11 @@ class BSplineCurve(Edge):
         Computes the abscissa of a 2D or 3D point using the least square
         method.
 
-        :param point: The point located on the B-spline curve
-        :type point: Union[:class:`volmdlr.Point2D`, :class:`volmdlr.Point3D`]
-        :param tol: The precision in terms of distance. Default value is 1e-4
-        :type tol: float, optional
-        :return: The abscissa of the point
+        :param point: The point located on the B-spline curve.
+        :type point: Union[:class:`volmdlr.Point2D`, :class:`volmdlr.Point3D`].
+        :param tol: The precision in terms of distance. Default value is 1e-4.
+        :type tol: float, optional.
+        :return: The abscissa of the point.
         :rtype: float
         """
         length = self.length()
@@ -731,6 +739,7 @@ class BSplineCurve(Edge):
 class Line2D(Line):
     """
     Define an infinite line given by two points.
+
     """
 
     def __init__(self, point1: volmdlr.Point2D,
@@ -744,38 +753,45 @@ class Line2D(Line):
 
     def rotation(self, center: volmdlr.Point2D, angle: float):
         """
-        Line2D rotation
-        :param center: rotation center
-        :param angle: angle rotation
-        :return: a new rotated Line2D
+        Line2D rotation.
+
+        :param center: rotation center.
+        :param angle: angle rotation.
+        :return: a new rotated Line2D.
         """
         return Line2D(*[point.rotation(center, angle)
                         for point in self.points])
 
     def rotation_inplace(self, center: volmdlr.Point2D, angle: float):
         """
-        Line2D rotation. Object is updated inplace
-        :param center: rotation center
-        :param angle: rotation angle
+        Line2D rotation. Object is updated inplace.
+
+        :param center: rotation center.
+        :param angle: rotation angle.
         """
         for point in self.points:
             point.rotation_inplace(center, angle)
 
     def translation(self, offset: volmdlr.Vector2D):
         """
-        Line2D translation
-        :param offset: translation vector
-        :return: A new translated Line2D
+        Line2D translation.
+
+        :param offset: translation vector.
+        :return: A new translated Line2D.
         """
         return Line2D(*[point.translation(offset) for point in self.points])
 
     def translation_inplace(self, offset: volmdlr.Vector2D):
         """
-        Line2D translation. Object is updated inplace
-        :param offset: translation vector
+        Line2D translation. Object is updated inplace.
+
+        :param offset: translation vector.
         """
         for point in self.points:
             point.translation_inplace(offset)
+
+    def frame_mapping(self, frame: volmdlr.Frame2D, side: str):
+        return Line2D(*[point.frame_mapping(frame, side) for point in self.points])
 
     def plot(self, ax=None, color='k', dashed=True):
         if ax is None:
@@ -828,6 +844,7 @@ class Line2D(Line):
         """
         Computes the two circles that are tangent to 2 lines and intersect
         a point located on one of the two lines.
+
         """
 
         # point will be called I(x_I, y_I)
@@ -967,9 +984,10 @@ class Line2D(Line):
 
     def point_distance(self, point2d):
         """
-        Calculates the distance of a line2d to a point2d
-        :param point2d: point to calculate distance
-        :return: distance to point
+        Calculates the distance of a line2d to a point2d.
+
+        :param point2d: point to calculate distance.
+        :return: distance to point.
         """
         vector_r = self.point1 - point2d
         vector_v = self.normal_vector()
@@ -979,7 +997,7 @@ class Line2D(Line):
 class BSplineCurve2D(BSplineCurve):
     """
     A class for 2 dimensional B-spline curves. The following rule must be
-    respected : `number of knots = number of control points + degree + 1`
+    respected : `number of knots = number of control points + degree + 1`.
 
     :param degree: The degree of the 2 dimensional B-spline curve
     :type degree: int
@@ -1022,10 +1040,12 @@ class BSplineCurve2D(BSplineCurve):
         """
         Computes the bounding rectangle of the 2 dimensional B-spline curve.
 
-        :return: The bounding rectangle
+        :return: The bounding rectangle.
         :rtype: :class:`volmdlr.core.BoundingRectangle`
         """
-        points = self.discretization_points()
+
+        points = self.discretization_points(number_points=5)
+
         points_x = [p.x for p in points]
         points_y = [p.y for p in points]
 
@@ -1163,7 +1183,8 @@ class BSplineCurve2D(BSplineCurve):
 
     def rotation(self, center: volmdlr.Point2D, angle: float):
         """
-        BSplineCurve2D rotation
+        BSplineCurve2D rotation.
+
         :param center: rotation center
         :param angle: angle rotation
         :return: a new rotated Line2D
@@ -1176,7 +1197,8 @@ class BSplineCurve2D(BSplineCurve):
 
     def rotation_inplace(self, center: volmdlr.Point2D, angle: float):
         """
-        BSplineCurve2D rotation. Object is updated inplace
+        BSplineCurve2D rotation. Object is updated inplace.
+
         :param center: rotation center
         :param angle: rotation angle
         """
@@ -1219,7 +1241,8 @@ class BSplineCurve2D(BSplineCurve):
 
     def to_wire(self, n: int):
         """
-        convert a bspline curve to a wire2d defined with 'n' line_segments
+        Convert a bspline curve to a wire2d defined with 'n' line_segments.
+
         """
 
         u = npy.linspace(0, 1, num=n + 1).tolist()
@@ -1232,7 +1255,8 @@ class BSplineCurve2D(BSplineCurve):
 
     def reverse(self):
         """
-        reverse the bspline's direction by reversing its start and end points
+        Reverse the bspline's direction by reversing its start and end points.
+
         """
 
         return self.__class__(degree=self.degree,
@@ -1254,13 +1278,20 @@ class BSplineCurve2D(BSplineCurve):
 
     def nearest_point_to(self, point):
         """
-        find out the nearest point on the linesegment to point
+        Find out the nearest point on the linesegment to point.
+
         """
 
         points = self.polygon_points(500)
         return point.nearest_point(points)
 
     def linesegment_intersections(self, linesegment):
+        """
+        Calculates intersections between a BSplineCurve2D and a LineSegment2D.
+
+        :param linesegment: linesegment to verify intersections.
+        :return: list with the intersections points.
+        """
         results = self.line_intersections(linesegment.to_line())
         intersections_points = []
         for result in results:
@@ -1270,7 +1301,8 @@ class BSplineCurve2D(BSplineCurve):
 
     def axial_symmetry(self, line):
         """
-        finds out the symmetric bsplinecurve2d according to a line
+        Finds out the symmetric bsplinecurve2d according to a line.
+
         """
 
         points_symmetry = [point.axial_symmetry(line) for point in self.control_points]
@@ -1308,12 +1340,14 @@ class BezierCurve2D(BSplineCurve2D):
 
 class LineSegment2D(LineSegment):
     """
-    Define a line segment limited by two points
+    Define a line segment limited by two points.
+
     """
 
     def __init__(self, start: volmdlr.Point2D, end: volmdlr.Point2D, *, name: str = ''):
         if start == end:
             raise NotImplementedError
+        self.points = [start, end]
         LineSegment.__init__(self, start, end, name=name)
 
     def __hash__(self):
@@ -1367,15 +1401,19 @@ class LineSegment2D(LineSegment):
     def straight_line_point_belongs(self, point):
         """
         Verifies if a point belongs to the surface created by closing the edge with a
-        line between its start and end points
-        :param point: Point to be verified
-        :return: Return True if the point belongs to this surface, or False otherwise
+        line between its start and end points.
+
+        :param point: Point to be verified.
+        :return: Return True if the point belongs to this surface, or False otherwise.
         """
         return self.point_belongs(point)
 
     def point_distance(self, point, return_other_point=False):
         """
-        Computes the distance of a point to segment of line
+        Computes the distance of a point to segment of line.
+
+        :param point: point to calculate distance.
+        :param return_other_points: Bool variable to return linesegment's corresponding point or not.
         """
         distance, point = volmdlr.LineSegment2DPointDistance(
             [(self.start.x, self.start.y), (self.end.x, self.end.y)],
@@ -1424,7 +1462,8 @@ class LineSegment2D(LineSegment):
 
     def linesegment_intersections(self, linesegment: 'LineSegment2D'):
         """
-        touching linesegments does not intersect
+        Touching linesegments does not intersect.
+
         """
         point = volmdlr.Point2D.line_intersection(self, linesegment)
         # TODO: May be these commented conditions should be used for linesegment_crossings
@@ -1508,7 +1547,8 @@ class LineSegment2D(LineSegment):
 
     def rotation(self, center: volmdlr.Point2D, angle: float):
         """
-        LineSegment2D rotation
+        LineSegment2D rotation.
+
         :param center: rotation center
         :param angle: angle rotation
         :return: a new rotated LineSegment2D
@@ -1518,34 +1558,38 @@ class LineSegment2D(LineSegment):
 
     def rotation_inplace(self, center: volmdlr.Point2D, angle: float):
         """
-        LineSegment2D rotation. Object is updated inplace
-        :param center: rotation center
-        :param angle: rotation angle
+        LineSegment2D rotation. Object is updated inplace.
+
+        :param center: rotation center.
+        :param angle: rotation angle.
         """
         for point in [self.start, self.end]:
             point.rotation_inplace(center, angle)
 
     def translation(self, offset: volmdlr.Vector2D):
         """
-        LineSegment2D translation
-        :param offset: translation vector
-        :return: A new translated LineSegment2D
+        LineSegment2D translation.
+
+        :param offset: translation vector.
+        :return: A new translated LineSegment2D.
         """
         return LineSegment2D(self.start.translation(offset),
                              self.end.translation(offset))
 
     def translation_inplace(self, offset: volmdlr.Vector2D):
         """
-        LineSegment2D translation. Object is updated inplace
-        :param offset: translation vector
+        LineSegment2D translation. Object is updated inplace.
+
+        :param offset: translation vector.
         """
         for point in [self.start, self.end]:
             point.translation_inplace(offset)
 
     def frame_mapping(self, frame: volmdlr.Frame2D, side: str):
         """
-        Changes vector frame_mapping and return a new LineSegment2D
-        side = 'old' or 'new'
+        Changes vector frame_mapping and return a new LineSegment2D.
+
+        side = 'old' or 'new'.
         """
         if side == 'old':
             new_start = frame.old_coordinates(self.start)
@@ -1559,8 +1603,10 @@ class LineSegment2D(LineSegment):
 
     def frame_mapping_inplace(self, frame: volmdlr.Frame2D, side: str):
         """
-        Changes vector frame_mapping and the object is updated inplace
-        side = 'old' or 'new'
+        Changes vector frame_mapping and the object is updated inplace.
+
+        :param frame: frame to execute the frame mapping.
+        :param side: 'old' or 'new'.
         """
         if side == 'old':
             new_start = frame.old_coordinates(self.start)
@@ -1591,9 +1637,8 @@ class LineSegment2D(LineSegment):
         return circle1, circle2
 
     def infinite_primitive(self, offset):
-        n = self.normal_vector()
+        n = self.unit_normal_vector()
         offset_point_1 = self.start + offset * n
-
         offset_point_2 = self.end + offset * n
 
         return Line2D(offset_point_1, offset_point_2)
@@ -1606,7 +1651,8 @@ class LineSegment2D(LineSegment):
 
     def to_wire(self, n: int):
         """
-        convert a linesegment2d to a wire2d defined with 'n' line_segments
+        Convert a linesegment2d to a wire2d defined with 'n' line_segments.
+
         """
 
         points = self.discretization_points(number_points=n + 1)
@@ -1614,7 +1660,8 @@ class LineSegment2D(LineSegment):
 
     def nearest_point_to(self, point):
         """
-        find out the nearest point on the linesegment to point
+        Find out the nearest point on the linesegment to point.
+
         """
 
         points = self.discretization_points(number_points=500)
@@ -1622,7 +1669,7 @@ class LineSegment2D(LineSegment):
 
     def axial_symmetry(self, line):
         """
-        finds out the symmetric linesegment2d according to a line
+        Finds out the symmetric linesegment2d according to a line.
         """
 
         points_symmetry = [point.axial_symmetry(line) for point in [self.start, self.end]]
@@ -1644,8 +1691,9 @@ class Arc(Edge):
     @property
     def center(self):
         """
-        Gets the arc's center
-        :return: The center of the arc
+        Gets the arc's center.
+
+        :return: The center of the arc.
         """
         raise NotImplementedError(
             'the property method center must be overloaded by subclassing'
@@ -1654,8 +1702,9 @@ class Arc(Edge):
     @property
     def angle(self):
         """
-        Gets the angle of the arc
-        :return: The angle of the arc
+        Gets the angle of the arc.
+
+        :return: The angle of the arc.
         """
         return NotImplementedError(
             'the property method angle must be overloaded by subclassing'
@@ -1664,8 +1713,9 @@ class Arc(Edge):
     @property
     def is_trigo(self):
         """
-        Verifies if arc is trigowise or clockwise
-        :return: True if trigowise or False otherwise
+        Verifies if arc is trigowise or clockwise.
+
+        :return: True if trigowise or False otherwise.
         """
         return NotImplementedError(
             'the property method is_trigo must be overloaded by subclassing'
@@ -1679,8 +1729,9 @@ class Arc(Edge):
 
     def length(self):
         """
-        Calculates the length of the Arc, with its radius and it arc angle
-        :return: the length fo the Arc
+        Calculates the length of the Arc, with its radius and it arc angle.
+
+        :return: the length fo the Arc.
         """
         return self.radius * abs(self.angle)
 
@@ -1695,10 +1746,11 @@ class Arc(Edge):
     @staticmethod
     def get_clockwise_and_trigowise_paths(radius_1, radius_2, radius_i):
         """
-        :param radius_1: radius from center to start point
-        :param radius_2: radius form center ro end point
-        :param radius_i: radius from center to interior point
-        :return: the clockwise and trigowise paths
+
+        :param radius_1: radius from center to start point.
+        :param radius_2: radius form center ro end point.
+        :param radius_i: radius from center to interior point.
+        :return: the clockwise and trigowise paths.
         """
         angle1 = math.atan2(radius_1.y, radius_1.x)
         anglei = math.atan2(radius_i.y, radius_i.x)
@@ -1730,7 +1782,8 @@ class Arc(Edge):
 
     def discretization_points(self, *, number_points: int = None, angle_resolution: int = None):
         """
-        discretize a Edge to have "n" points
+        Discretize a Edge to have "n" points.
+
         :param number_points: the number of points (including start and end points)
              if unset, only start and end will be returned
         :param angle_resolution: if set, the sampling will be adapted to have a controlled angular distance. Usefull
@@ -1866,7 +1919,8 @@ class Arc2D(Arc):
 
     def point_belongs(self, point2d, abs_tol=1e-10):
         """
-        check if a Point2D belongs to the Arc2D
+        Check if a Point2D belongs to the Arc2D.
+
         """
         vector_start = self.start - self.center
         vector_end = self.end - self.center
@@ -2107,9 +2161,10 @@ class Arc2D(Arc):
     def straight_line_point_belongs(self, point):
         """
         Verifies if a point belongs to the surface created by closing the edge with a
-        line between its start and end points
-        :param point_2d: Point to be verified
-        :return: Return True if the point belongs to this surface, or False otherwise
+        line between its start and end points.
+
+        :param point_2d: Point to be verified.
+        :return: Return True if the point belongs to this surface, or False otherwise.
         """
         if self.point_belongs(point_2d):
             return True
@@ -2151,19 +2206,21 @@ class Arc2D(Arc):
 
     def rotation(self, center: volmdlr.Point2D, angle: float):
         """
-        Arc2D rotation
+        Arc2D rotation.
+
         :param center: rotation center
-        :param angle: angle rotation
-        :return: a new rotated Arc2D
+        :param angle: angle rotation.
+        :return: a new rotated Arc2D.
         """
         return Arc2D(*[point.rotation(center, angle,) for point in
                        [self.start, self.interior, self.end]])
 
     def rotation_inplace(self, center: volmdlr.Point2D, angle: float):
         """
-        Arc2D rotation. Object is updated inplace
-        :param center: rotation center
-        :param angle: rotation angle
+        Arc2D rotation. Object is updated inplace.
+
+        :param center: rotation center.
+        :param angle: rotation angle.
         """
         self.start.rotation_inplace(center, angle)
         self.interior.rotation_inplace(center, angle)
@@ -2175,17 +2232,19 @@ class Arc2D(Arc):
 
     def translation(self, offset: volmdlr.Vector2D):
         """
-        Arc2D translation
-        :param offset: translation vector
-        :return: A new translated Arc2D
+        Arc2D translation.
+
+        :param offset: translation vector.
+        :return: A new translated Arc2D.
         """
         return Arc2D(*[point.translation(offset) for point in
                        [self.start, self.interior, self.end]])
 
     def translation_inplace(self, offset: volmdlr.Vector2D):
         """
-        Arc2D translation. Object is updated inplace
-        :param offset: translation vector
+        Arc2D translation. Object is updated inplace.
+
+        :param offset: translation vector.
         """
         self.start.translation_inplace(offset)
         self.interior.translation_inplace(offset)
@@ -2197,7 +2256,8 @@ class Arc2D(Arc):
 
     def frame_mapping(self, frame: volmdlr.Frame2D, side: str):
         """
-        Changes vector frame_mapping and return a new Arc2D
+        Changes vector frame_mapping and return a new Arc2D.
+
         side = 'old' or 'new'
         """
         return Arc2D(*[point.frame_mapping(frame, side) for point in
@@ -2205,7 +2265,7 @@ class Arc2D(Arc):
 
     def frame_mapping_inplace(self, frame: volmdlr.Frame2D, side: str):
         """
-        Changes vector frame_mapping and the object is updated inplace
+        Changes vector frame_mapping and the object is updated inplace.
         side = 'old' or 'new'
         """
         self.__init__(*[point.frame_mapping(frame, side) for point in
@@ -2213,7 +2273,8 @@ class Arc2D(Arc):
 
     def second_moment_area(self, point):
         """
-        Second moment area of part of disk
+        Second moment area of part of disk.
+
         """
         if self.angle2 < self.angle1:
             angle2 = self.angle2 + volmdlr.TWO_PI
@@ -2285,14 +2346,16 @@ class Arc2D(Arc):
 
     def to_wire(self, angle_resolution: float = 10.):
         """
-        convert an arc to a wire2d defined with line_segments
+        convert an arc to a wire2d defined with line_segments.
+
         """
 
         return volmdlr.wires.Wire2D.from_points(self.polygon_points(angle_resolution))
 
     def axial_symmetry(self, line):
         """
-        finds out the symmetric arc2d according to a line
+        Finds out the symmetric arc2d according to a line.
+
         """
 
         points_symmetry = [point.axial_symmetry(line) for point in [self.start, self.interior, self.end]]
@@ -2305,7 +2368,8 @@ class Arc2D(Arc):
 class FullArc2D(Arc2D):
     """
     An edge that starts at start_end, ends at the same point after having described
-    a circle
+    a circle.
+
     """
 
     def __init__(self, center: volmdlr.Point2D, start_end: volmdlr.Point2D,
@@ -2370,7 +2434,8 @@ class FullArc2D(Arc2D):
     def straight_line_point_belongs(self, point):
         """
         Verifies if a point belongs to the surface created by closing the edge with a
-        line between its start and end points
+        line between its start and end points.
+
         :param point2d: Point to be verified
         :return: Return True if the point belongs to this surface, or False otherwise
         """
@@ -2655,7 +2720,8 @@ class ArcEllipse2D(Edge):
 
     def length(self):
         """
-        Calculates the length of the arcellipse2d
+        Calculates the length of the arcellipse2d.
+
         :return: arcellipse2d's length
         """
         length = self.abscissa(self.end)
@@ -2663,7 +2729,8 @@ class ArcEllipse2D(Edge):
 
     def point_belongs(self, point, abs_tol: float = 1e-6):
         """
-        Verifies if a point belongs to the arcellipse2d
+        Verifies if a point belongs to the arcellipse2d.
+
         :param point: point to be verified
         :param abs_tol: tolerance applied during calculations
         :return: True if the point belongs, False otherwise
@@ -2684,7 +2751,8 @@ class ArcEllipse2D(Edge):
 
     def abscissa(self, point: volmdlr.Point2D):
         """
-        Calculates the abscissa of a given point
+        Calculates the abscissa of a given point.
+
         :param point: point for calculating abscissa
         :return: a float, between 0 and the arcellise2d's lenght
         """
@@ -2706,8 +2774,9 @@ class ArcEllipse2D(Edge):
 
     def bounding_rectangle(self):
         """
-        Calculates the bounding rectangle for the arcellipse2d
-        :return: volmdlr.core.BoudingRectangle object
+        Calculates the bounding rectangle for the arcellipse2d.
+
+        :return: volmdlr.core.BoudingRectangle object.
         """
         min_a, max_a = self.center - self.Gradius * self.major_dir, self.center + self.Gradius * self.major_dir
         min_b, max_b = self.center - self.Sradius * self.minor_dir, self.center + self.Sradius * self.minor_dir
@@ -2717,8 +2786,9 @@ class ArcEllipse2D(Edge):
 
     def straight_line_area(self):
         """
-        Calculates the area of the elliptic arc, with line drwan from start to end
-        :return: straight_line_area
+        Calculates the area of the elliptic arc, with line drwan from start to end.
+
+        :return: straight_line_area.
         """
         if self.angle >= math.pi:
             angle = volmdlr.TWO_PI - self.angle
@@ -2734,12 +2804,13 @@ class ArcEllipse2D(Edge):
 
     def discretization_points(self, *, number_points: int = None, angle_resolution: int = None):
         """
-        discretize an Edge to have "n" points
+        discretize an Edge to have "n" points.
+
         :param number_points: the number of points (including start and end points)
-             if unset, only start and end will be returned
+             if unset, only start and end will be returned.
         :param angle_resolution: if set, the sampling will be adapted to have a controlled angular distance. Usefull
-            to mesh an arc
-        :return: a list of sampled points
+            to mesh an arc.
+        :return: a list of sampled points.
         """
         if not number_points:
             if not angle_resolution:
@@ -2824,11 +2895,11 @@ class ArcEllipse2D(Edge):
 
 
 class Line3D(Line):
-    _non_eq_attributes = ['name', 'basis_primitives', 'bounding_box']
+    """
+    Define an infinite line passing through the 2 points.
 
     """
-    Define an infinite line passing through the 2 points
-    """
+    _non_eq_attributes = ['name', 'basis_primitives', 'bounding_box']
 
     def __init__(self, point1: volmdlr.Point3D, point2: volmdlr.Point3D,
                  name: str = ''):
@@ -2878,9 +2949,10 @@ class Line3D(Line):
 
     def line_distance(self, line2):
         """
-        Calculates the distance between two Line3D
-        :param line2: other Line3D
-        :return: The distance between the two lines
+        Calculates the distance between two Line3D.
+
+        :param line2: other Line3D.
+        :return: The distance between the two lines.
         """
         direction_vector1 = self.direction_vector()
         direction_vector2 = line2.direction_vector()
@@ -2893,9 +2965,10 @@ class Line3D(Line):
 
     def skew_to(self, line):
         """
-        Verifies if two Line3D are skew to each other, that is, they are not parallel and never intersect
-        :param line: othe line
-        :return: True if they are skew, False otherwise
+        Verifies if two Line3D are skew to each other, that is, they are not parallel and never intersect.
+
+        :param line: othe line.
+        :return: True if they are skew, False otherwise.
         """
         if self.direction_vector().is_colinear_to(line.direction_vector()):
             return False
@@ -3131,7 +3204,7 @@ class Line3D(Line):
         m, n, p = direction_vector2.x, direction_vector2.y, direction_vector2.z
         vector_components = [a, b, c, m, n, p]
         for i, component in enumerate(vector_components):
-            if abs(component) <= 1e-6:
+            if abs(component) <= 1e-5:
                 vector_components[i] = 0.0
         a, b, c, m, n, p = vector_components
         # if a == m == 0 and x2 != x1:
@@ -3149,7 +3222,7 @@ class Line3D(Line):
             coefficient_t = (b * (x2 - x1) - a * (y2 - y1)) / (n * a - b * m)
             coefficient_s = (n * (x2 - x1) - m * (y2 - y1)) / (n * a - b * m)
         elif a == m == 0:
-            if math.isclose(x2, x1, abs_tol=1e-6) and c * n != b * p:
+            if math.isclose(x2, x1, abs_tol=1e-5) and c * n != b * p:
                 if b != 0:
                     coefficient_t = ((z2 - z1) * b - c * (y2 - y1)) / (c * n - b * p)
                     coefficient_s = ((y2 - y1) + n * coefficient_t) / b
@@ -3159,7 +3232,7 @@ class Line3D(Line):
             else:
                 raise NotImplementedError
         elif b == n == 0.:
-            if math.isclose(y2, y1, abs_tol=1e-6) and c * m != a * p:
+            if math.isclose(y2, y1, abs_tol=1e-5) and c * m != a * p:
                 if a != 0.:
                     coefficient_t = ((z2 - z1) * a - c * (x2 - x1)) / (c * m - a * p)
                     coefficient_s = ((x2 - x1) + m * coefficient_t) / a
@@ -3199,8 +3272,16 @@ class Line3D(Line):
         content += f"#{current_id} = LINE('{self.name}',#{p1_id},#{u_id});\n"
         return content, current_id
 
-    def to_2d(self, plane_origin, x1, x2):
-        p2d = [p.to_2d(plane_origin, x1, x2) for p in (self.point1, self.point2)]
+    def to_2d(self, plane_origin, x, y):
+        """
+        Tranforms a Line3D into an Line2D, given an plane origin and a u and v plane vector.
+
+        :param plane_origin: plane origin.
+        :param x: plane u vector.
+        :param y: plane v vector.
+        :return: Line2D.
+        """
+        p2d = [p.to_2d(plane_origin, x, y) for p in (self.point1, self.point2)]
         if p2d[0] == p2d[1]:
             return None
         return Line2D(*p2d, name=self.name)
@@ -3283,8 +3364,10 @@ class LineSegment3D(LineSegment):
         return distance
 
     def plane_projection2d(self, center, x, y):
-        return LineSegment2D(self.start.plane_projection2d(center, x, y),
-                             self.end.plane_projection2d(center, x, y))
+        start, end = self.start.plane_projection2d(center, x, y), self.end.plane_projection2d(center, x, y)
+        if start != end:
+            return LineSegment2D(start, end)
+        return None
 
     # def intersection(self, segment2):
     #     x1 = self.start.x
@@ -3545,8 +3628,16 @@ class LineSegment3D(LineSegment):
 
         return s
 
-    def to_2d(self, plane_origin, x1, x2):
-        p2d = [p.to_2d(plane_origin, x1, x2) for p in (self.start, self.end)]
+    def to_2d(self, plane_origin, x, y):
+        """
+        Tranforms a LineSegment3D into an LineSegment2D, given an plane origin and a u and v plane vector.
+
+        :param plane_origin: plane origin.
+        :param x: plane u vector.
+        :param y: plane v vector.
+        :return: LineSegment2D.
+        """
+        p2d = [p.to_2d(plane_origin, x, y) for p in (self.start, self.end)]
         if p2d[0] == p2d[1]:
             return None
         return LineSegment2D(*p2d, name=self.name)
@@ -4252,8 +4343,16 @@ class BSplineCurve3D(BSplineCurve, volmdlr.core.Primitive3D):
             ax.plot(x, y, z, 'o', color=color, alpha=alpha)
         return ax
 
-    def to_2d(self, plane_origin, x1, x2):
-        control_points2d = [p.to_2d(plane_origin, x1, x2) for p in
+    def to_2d(self, plane_origin, x, y):
+        """
+        Tranforms a BSplineCurve3D into an BSplineCurve2D, given an plane origin and a u and v plane vector.
+
+        :param plane_origin: plane origin.
+        :param x: plane u vector.
+        :param y: plane v vector.
+        :return: BSplineCurve2D.
+        """
+        control_points2d = [p.to_2d(plane_origin, x, y) for p in
                             self.control_points]
         return BSplineCurve2D(self.degree, control_points2d,
                               self.knot_multiplicities, self.knots,
@@ -4394,6 +4493,11 @@ class Arc3D(Arc):
         self._bbox = new_bounding_box
 
     def get_bounding_box(self):
+        """
+        Calculates the bounding box of the Arc3D.
+
+        :return: a volmdlr.core.BoundingBox object.
+        """
         # TODO: implement exact calculation
 
         points = self.polygon_points()
@@ -4552,31 +4656,65 @@ class Arc3D(Arc):
         return [self.start, self.interior, self.end]
 
     def reverse(self):
+        """
+        Defines a new Arc3D, odentical to self, but in the oposite direction.
+
+        """
         return self.__class__(self.end.copy(),
                               self.interior.copy(),
                               self.start.copy())
 
     def point_at_abscissa(self, abscissa):
+        """
+        Calculates a point in the Arc2D at a given abscissa.
+
+        :param abscissa: abscissa where in the curve the point should be calculated.
+        :return: Corresponding point.
+        """
         return self.start.rotation(self.center, self.normal,
                                    abscissa / self.radius)
 
     def normal_vector(self, abscissa):
+        """
+        Calculates a normal vector at a given abscissa of the Arc3D.
+
+        :param abscissa: abscissa where in the curve the normal vector should be calculated.
+        :return: Corresponding normal vector.
+        """
         theta = abscissa / self.radius
         n_0 = self.center - self.start
         normal = n_0.rotation(self.center, self.normal, theta)
         return normal
 
     def unit_normal_vector(self, abscissa):
+        """
+        Calculates a unit normal vector at a given abscissa of the Arc3D.
+
+        :param abscissa: abscissa where in the curve the unit normal vector should be calculated.
+        :return: Corresponding unit normal vector.
+        """
         normal_vector = self.normal_vector(abscissa)
         normal_vector.normalize()
         return normal_vector
 
     def direction_vector(self, abscissa):
+        """
+        Calculates a direction vector at a given abscissa of the Arc3D.
+
+        :param abscissa: abscissa where in the curve the direction vector should be calculated.
+        :return: Corresponding direction vector.
+        """
         normal_vector = self.normal_vector(abscissa)
         tangent = normal_vector.cross(self.normal)
         return tangent
 
     def unit_direction_vector(self, abscissa):
+        """
+        Calculates a unit direction vector at a given abscissa of the Arc3D.
+
+        :param abscissa: abscissa where in the curve the unit direction vector should be calculated.
+        :return: Corresponding unit direction vector.
+        """
         direction_vector = self.direction_vector(abscissa)
         direction_vector.normalize()
         return direction_vector
@@ -4584,7 +4722,8 @@ class Arc3D(Arc):
     def rotation(self, center: volmdlr.Point3D,
                  axis: volmdlr.Vector3D, angle: float):
         """
-        Arc3D rotation
+        Arc3D rotation.
+
         :param center: rotation center
         :param axis: rotation axis
         :param angle: angle rotation
@@ -4598,7 +4737,8 @@ class Arc3D(Arc):
     def rotation_inplace(self, center: volmdlr.Point3D,
                          axis: volmdlr.Vector3D, angle: float):
         """
-        Arc3D rotation. Object is updated inplace
+        Arc3D rotation. Object is updated inplace.
+
         :param center: rotation center
         :param axis: rotation axis
         :param angle: rotation angle
@@ -4614,7 +4754,8 @@ class Arc3D(Arc):
 
     def translation(self, offset: volmdlr.Vector3D):
         """
-        Arc3D translation
+        Arc3D translation.
+
         :param offset: translation vector
         :return: A new translated Arc3D
         """
@@ -4625,8 +4766,9 @@ class Arc3D(Arc):
 
     def translation_inplace(self, offset: volmdlr.Vector3D):
         """
-        Arc3D translation. Object is updated inplace
-        :param offset: translation vector
+        Arc3D translation. Object is updated inplace.
+
+        :param offset: translation vector.
         """
         self.center.translation_inplace(offset)
         self.start.translation_inplace(offset)
@@ -4721,7 +4863,8 @@ class Arc3D(Arc):
 
     def frame_mapping(self, frame: volmdlr.Frame3D, side: str):
         """
-        Changes vector frame_mapping and return a new Arc3D
+        Changes vector frame_mapping and return a new Arc3D.
+
         side = 'old' or 'new'
         """
         new_start, new_interior, new_end =\
@@ -4731,7 +4874,8 @@ class Arc3D(Arc):
 
     def frame_mapping_inplace(self, frame: volmdlr.Frame3D, side: str):
         """
-        Changes vector frame_mapping and the object is updated inplace
+        Changes vector frame_mapping and the object is updated inplace.
+
         side = 'old' or 'new'
         """
         new_start, new_interior, new_end = \
@@ -4739,6 +4883,12 @@ class Arc3D(Arc):
         self.start, self.interior, self.end = new_start, new_interior, new_end
 
     def abscissa(self, point3d: volmdlr.Point3D):
+        """
+        Calculates the abscissa given a point in the Arc3D.
+
+        :param point3d: point to calculate the abscissa.
+        :return: corresponding abscissa.
+        """
         x, y, _ = self.frame.new_coordinates(point3d)
         u1 = x / self.radius
         u2 = y / self.radius
@@ -4747,6 +4897,12 @@ class Arc3D(Arc):
         return self.radius * abs(theta)
 
     def split(self, split_point: volmdlr.Point3D):
+        """
+        Splits the Arc2D in two at a given point.
+
+        :param split_point: splitting point
+        :return: two Arc2D.
+        """
         abscissa = self.abscissa(split_point)
 
         return [Arc3D(self.start,
@@ -4758,6 +4914,14 @@ class Arc3D(Arc):
                 ]
 
     def to_2d(self, plane_origin, x, y):
+        """
+        Tranforms a Arc3D into an Arc2D, given an plane origin and a u and v plane vector.
+
+        :param plane_origin: plane origin.
+        :param x: plane u vector.
+        :param y: plane v vector.
+        :return: Arc2D.
+        """
         ps = self.start.to_2d(plane_origin, x, y)
         pi = self.interior.to_2d(plane_origin, x, y)
         pe = self.end.to_2d(plane_origin, x, y)
@@ -4998,22 +5162,25 @@ class Arc3D(Arc):
             start_id, end_id, curve_id)
         return content, [current_id]
 
-    def point_belongs(self, point3d, abs_tol=1e-10):
+    def point_belongs(self, point3d, abs_tol: float = 1e-6):
         """
-        check if a point3d belongs to the arc_3d or not
-        """
-        def f(x):
-            return (point3d - self.point_at_abscissa(x)).norm()
-        length_ = self.length()
-        x = npy.linspace(0, length_, 5)
-        x_init = []
-        for xi in x:
-            x_init.append(xi)
+        Check if a point3d belongs to the arc_3d or not.
 
-        for x0 in x_init:
-            z = scp.optimize.least_squares(f, x0=x0, bounds=([0, length_]))
-            if z.fun < abs_tol:
-                return True
+        :param point3d: point to be verified is on arc
+        :return: True if point is on Arc, False otherwise.
+        """
+        if not math.isclose(point3d.point_distance(self.center), self.radius, abs_tol=abs_tol):
+            return False
+        vector1 = self.start - self.center
+        vector2 = self.interior - self.center
+        vector3 = point3d - self.center
+        if not math.isclose(vector1.dot(vector2.cross(vector3)), 0.0, abs_tol=abs_tol):
+            return False
+        point_abscissa = self.abscissa(point3d)
+        abscissa_start = self.abscissa(self.start)
+        abscissa_end = self.abscissa(self.end)
+        if abscissa_start <= point_abscissa <= abscissa_end:
+            return True
         return False
 
     def triangulation(self):
@@ -5021,6 +5188,34 @@ class Arc3D(Arc):
 
     def middle_point(self):
         return self.point_at_abscissa(self.length() / 2)
+
+    def line_intersections(self, line3d: Line3D):
+        """
+        Calculates intersections between an Arc3D and a Line3D.
+
+        :param linesegment3d: linesegment to verify intersections.
+        :return: list with intersections points between line and Arc3D.
+        """
+        circle3d_lineseg_inters = vm_utils_intersections.circle_3d_linesegment_intersections(self, line3d)
+        linesegment_intersections = []
+        for intersection in circle3d_lineseg_inters:
+            if self.point_belongs(intersection, 1e-6):
+                linesegment_intersections.append(intersection)
+        return linesegment_intersections
+
+    def linesegment_intersections(self, linesegment3d: LineSegment3D):
+        """
+        Calculates intersections between an Arc3D and a LineSegment3D.
+
+        :param linesegment3d: linesegment to verify intersections.
+        :return: list with intersections points between linesegment and Arc3D.
+        """
+        linesegment_intersections = []
+        intersections = self.line_intersections(linesegment3d)
+        for intersection in intersections:
+            if linesegment3d.point_belongs(intersection):
+                linesegment_intersections.append(intersection)
+        return linesegment_intersections
 
 
 class FullArc3D(Arc3D):
@@ -5064,9 +5259,17 @@ class FullArc3D(Arc3D):
     def copy(self, *args, **kwargs):
         return FullArc3D(self._center.copy(), self.end.copy(), self._normal.copy())
 
-    def to_2d(self, plane_origin, x1, x2):
-        center = self.center.to_2d(plane_origin, x1, x2)
-        start_end = self.start.to_2d(plane_origin, x1, x2)
+    def to_2d(self, plane_origin, x, y):
+        """
+        Tranforms a FullArc3D into an FullArc2D, given an plane origin and a u and v plane vector.
+
+        :param plane_origin: plane origin.
+        :param x: plane u vector.
+        :param y: plane v vector.
+        :return: FullArc2D.
+        """
+        center = self.center.to_2d(plane_origin, x, y)
+        start_end = self.start.to_2d(plane_origin, x, y)
         return FullArc2D(center, start_end)
 
     def to_step(self, current_id, surface_id=None):
@@ -5393,6 +5596,14 @@ class ArcEllipse3D(Edge):
     points = property(_get_points)
 
     def to_2d(self, plane_origin, x, y):
+        """
+        Tranforms a ArcEllipse3D into an ArcEllipse2D, given an plane origin and a u and v plane vector.
+
+        :param plane_origin: plane origin.
+        :param x: plane u vector.
+        :param y: plane v vector.
+        :return: ArcEllipse2D.
+        """
         point_start2d = self.start.to_2d(plane_origin, x, y)
         point_interior2d = self.interior.to_2d(plane_origin, x, y)
         point_end2d = self.end.to_2d(plane_origin, x, y)